LayoutTests/fast/events/mouseenter-mouseleave-expected.txt

mouseover on outer1
mouseout on outer1
mouseover on inner1
mouseover on inner1
mouseout on inner1
mouseout on inner1
mouseover on inner3
mouseover on inner3
mouseover on inner3
mouseout on inner3
mouseout on inner3
mouseout on inner3
mouseover on outer1
mouseout on outer1

mouseenter on outer1
mouseenter on inner1
mouseleave on inner1
mouseenter on inner3
mouseenter on inner2
mouseleave on inner3
mouseleave on inner2
mouseleave on outer1
